카드 순서 변경 기능 추가

+++ src/main/java/com/sparta/tse/domain/card/dto/request/CardRequestDto.java
@@ -15,14 +15,14 @@ public class CardRequestDto {
     public LocalDateTime startAt;
     public LocalDateTime endAt;
     public List<Long> userId;
-    public int cardSequence;
+    public Long cardSequence;
 
-    public void cardSequence(int maxSequence) {
+    public void cardSequence(Long maxSequence) {
         this.cardSequence = maxSequence;
     }
 
     public CardRequestDto (Long listId, String cardTitle, String cardContent, LocalDateTime startAt,
-                           LocalDateTime endAt, List<Long> userId, int cardSequence) {
+                           LocalDateTime endAt, List<Long> userId, Long cardSequence) {
         this.listId = listId;
         this.cardTitle = cardTitle;
         this.cardContent = cardContent;
